requirements.txt
flask
opencv-python
gunicorn
pdf2image
python-dotenv